refactor: organize opportunity API into views/ and serializers/ subpackages
Split opportunity/api/views.py into views/{mobile,automation}.py and
opportunity/api/serializers.py into serializers/{mobile,automation}.py.
The existing mobile-facing API lives under the mobile modules; the new
automation API lives under the automation modules.
