+++ commcare_connect/data_export/serializer.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
-from commcare_connect.opportunity.api.serializers import (
+from commcare_connect.opportunity.api.serializers.mobile import (
     CommCareAppSerializer,
     OpportunityClaimLimitSerializer,
     OpportunityVerificationFlagsSerializer,
