+++ commcare_connect/opportunity/api/views/automation.py
@@ -5,7 +5,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from commcare_connect.opportunity.api.automation_serializers import (
+from commcare_connect.opportunity.api.serializers.automation import (
     OpportunityActivateResponseSerializer,
     PaymentUnitListCreateSerializer,
     PaymentUnitResponseSerializer,
