+++ commcare_connect/opportunity/views.py
@@ -59,7 +59,7 @@ from waffle import switch_is_active
 from commcare_connect.connect_id_client import fetch_users
 from commcare_connect.flags.switch_names import INVOICE_REVIEW, UPDATES_TO_MARK_AS_PAID_WORKFLOW, WORKER_VISITS_TASKS
 from commcare_connect.form_receiver.serializers import XFormSerializer
-from commcare_connect.opportunity.api.serializers import remove_opportunity_access_cache
+from commcare_connect.opportunity.api.serializers.mobile import remove_opportunity_access_cache
 from commcare_connect.opportunity.app_xml import AppNoBuildException
 from commcare_connect.opportunity.filters import (
     DeliverFilterSet,
